--- app/services/dataset/dataset_deletion_service.py
@@ -1,81 +1,0 @@
-from app.config import settings
-from app.logging_config import logging
-from app.models.collection_exericise_end_data import CollectionExerciseEndData
-from app.models.dataset_models import DatasetMetadata
-from app.models.deletion_models import DeleteMetadata
-from app.repositories.dataset_deletion.firestore_dataset_deletion_repository import (
-    DeletionMetadataFirebaseRepository,
-)
-from app.util.firebase_loader import FirebaseLoader
-from app.services.dataset.dataset_service import DatasetService
-from app.services.shared.datetime_service import DatetimeService
-
-logger = logging.getLogger(__name__)
-
-
-class DatasetDeletionService:
-
-    def __init__(self, firebase_loader: FirebaseLoader) -> None:
-        self.delete_repository = DeletionMetadataFirebaseRepository(firebase_loader)
-        self.dataset_service = DatasetService(firebase_loader)
-
-    def process_collection_exercise_end_message(
-        self, collection_exercise_end_data: CollectionExerciseEndData
-    ):
-        # todo: validation on period_id and survey_id once we have go the 'real' message
-        collection_has_dataset_guid = (
-            self._check_if_collection_has_dataset_guid(
-                collection_exercise_end_data
-            )
-        )
-
-        if collection_has_dataset_guid:
-            list_dataset_metadata = self._collect_metadata_for_period_and_survey(
-                collection_exercise_end_data
-            )
-            self._mark_collections_for_deletion(list_dataset_metadata)
-        else:
-            logger.debug("Supplementary data not data found")
-
-    def _check_if_collection_has_dataset_guid(
-        self, collection_exercise_end_data: CollectionExerciseEndData
-    ) -> bool:
-        if (
-            collection_exercise_end_data.dataset_guid is None
-            or collection_exercise_end_data.dataset_guid == ""
-        ):
-            supplementary_data_available = False
-        else:
-            supplementary_data_available = True
-        return supplementary_data_available
-
-
-    def _collect_metadata_for_period_and_survey(
-        self, collection_exercise_end_data: CollectionExerciseEndData
-    ) -> list[DatasetMetadata]:
-        logger.info("Collecting all dataset versions for period and survey")
-        logger.info(f"Collecting all dataset versions for survey_id: {collection_exercise_end_data.survey_id} and period_id: {collection_exercise_end_data.period_id}")
-        return self.dataset_service.get_dataset_metadata_collection(
-            collection_exercise_end_data.survey_id, collection_exercise_end_data.period_id
-        )
-
-
-    def _mark_collections_for_deletion(
-        self, list_dataset_metadata: list[DatasetMetadata]
-    ):
-        time_now = DatetimeService.get_current_date_and_time().strftime(settings.TIME_FORMAT)
-        for dataset_metadata in list_dataset_metadata:
-            logger.debug(f"Dataset_metadata {dataset_metadata}")
-            delete_metadata: DeleteMetadata = DeleteMetadata(
-                **{
-                    "dataset_guid": dataset_metadata.dataset_id,
-                    "period_id": dataset_metadata.period_id,
-                    "survey_id": dataset_metadata.survey_id,
-                    "sds_dataset_version": dataset_metadata.sds_dataset_version,
-                    "status": "Pending",
-                    "mark_deleted_at": time_now,
-                    "deleted_at": "n/a",
-                }
-            )
-            logger.debug(f"Marking dataset for deletion {delete_metadata}")
-            self.delete_repository.mark_dataset_for_deletion(delete_metadata)
